Add user prompt to ask for conformation

+++ actions.py
@@ -2,8 +2,22 @@ from label_diff import LabelDiff
 import github_api
 
 
-def applyAllCreate(diff: LabelDiff, report=None):
+def _confirm():
+    while True:
+        selection = input("Is this ok [Y/n]: ").lower()
+        if selection == "n":
+            return False
+        elif selection == "y" or selection == '':
+            return True
+
+
+def applyAllCreate(diff: LabelDiff, yes: bool = False, report=None):
     for label in diff.missing:
+        if report is not None:
+            report(diff, 'create', label)
+        if not yes:
+            if not _confirm():
+                continue
         github_api.createLabel(
             diff.namespace,
             diff.repository,
@@ -11,19 +25,25 @@ def applyAllCreate(diff: LabelDiff, report=None):
             label['description'],
             label['color']
         )
-        if report is not None:
-            report(diff, 'create', label)
 
 
-def applyAllDelete(diff: LabelDiff, report=None):
+def applyAllDelete(diff: LabelDiff, yes: bool = False, report=None):
     for label in diff.extra:
-        github_api.deleteLabel(diff.namespace, diff.repository, label['name'])
         if report is not None:
             report(diff, 'delete', label)
+        if not yes:
+            if not _confirm():
+                continue
+        github_api.deleteLabel(diff.namespace, diff.repository, label['name'])
 
 
-def applyAllModify(diff: LabelDiff, report=None):
+def applyAllModify(diff: LabelDiff, yes: bool = False, report=None):
     for label in diff.diff:
+        if report is not None:
+            report(diff, 'modify', label)
+        if not yes:
+            if not _confirm():
+                continue
         github_api.updateLabel(
             diff.namespace,
             diff.repository,
@@ -31,5 +51,3 @@ def applyAllModify(diff: LabelDiff, report=None):
             description=label['truth']['description'] if 'description' in label['delta'] else None,
             color=label['truth']['color'] if 'color' in label['delta'] else None
         )
-        if report is not None:
-            report(diff, 'modify', label)
